+++ packages/core/src/button/private/types/ButtonSizesProps.ts
@@ -1,0 +1,15 @@
+import Size from '../../../private/enums/Size';
+
+type ButtonSizesProps =
+  | Size.XSMALL
+  | Size.SMALL
+  | Size.MEDIUM
+  | Size.LARGE
+  | Size.XLARGE
+  | 'xsmall'
+  | 'small'
+  | 'medium'
+  | 'large'
+  | 'xlarge';
+
+export default ButtonSizesProps;