+++ packages/core/src/checkbox/Checkbox.tsx
@@ -2,20 +2,7 @@ import React, { forwardRef, useState } from 'react';
 import CheckboxIconWrapper from './styles/CheckboxIconWrapper';
 import CheckboxInput from './styles/CheckboxInput';
 import CheckboxLabel from './styles/CheckboxLabel';
-import type Variants from './private/Variants';
-
-export interface CheckboxProps {
-  disabled?: boolean;
-  ariaLabel?: string;
-  label?: JSX.Element | string;
-  id?: string;
-  variant?: Variants;
-  checked?: boolean;
-  onClick?: (event?: React.MouseEvent<HTMLInputElement>) => void;
-  onChange?: (event: React.ChangeEvent<HTMLInputElement>) => void;
-  dir?: 'ltr' | 'rtl' | 'auto';
-  readOnly?: boolean;
-}
+import type CheckboxProps from './private/types/CheckboxProps';
 
 const Checkbox = forwardRef<HTMLInputElement, CheckboxProps>(
   (
@@ -23,7 +10,6 @@ const Checkbox = forwardRef<HTMLInputElement, CheckboxProps>(
     ref
   ) => {
     const [isChecked, setIsChecked] = useState(false);
-
     return (
       <CheckboxLabel htmlFor={id} variant={variant}>
         <CheckboxInput