+++ packages/core/src/checkbox/private/Variants.ts
@@ -1,3 +1,0 @@
-type Variants = 'option';
-
-export default Variants;