+++ next-docs/CHANGELOG.md
@@ -1,5 +1,80 @@
 # Change Log
 
+## 9.0.1
+
+### Patch Changes
+
+- Bump corrupted 9.0.0
+- Updated dependencies [undefined]
+  - @heathmont/moon-accordion@9.0.1
+  - @heathmont/moon-assets@9.0.1
+  - @heathmont/moon-charts@9.0.1
+  - @heathmont/moon-components@9.0.1
+  - @heathmont/moon-core@9.0.1
+  - @heathmont/moon-datepicker@9.0.1
+  - @heathmont/moon-draggabletable@9.0.1
+  - @heathmont/moon-icons@9.0.1
+  - @heathmont/moon-pagination@9.0.1
+  - @heathmont/moon-popover@9.0.1
+  - @heathmont/moon-select@9.0.1
+  - @heathmont/moon-sidebar@9.0.1
+  - @heathmont/moon-table@9.0.1
+  - @heathmont/moon-themes@9.0.1
+  - @heathmont/moon-utils@9.0.1
+  - @heathmont/moon-core-tw@9.0.1
+
+## 9.0.0
+
+### Major Changes
+
+- React 18
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-accordion@9.0.0
+  - @heathmont/moon-assets@9.0.0
+  - @heathmont/moon-charts@9.0.0
+  - @heathmont/moon-components@9.0.0
+  - @heathmont/moon-core@9.0.0
+  - @heathmont/moon-datepicker@9.0.0
+  - @heathmont/moon-draggabletable@9.0.0
+  - @heathmont/moon-icons@9.0.0
+  - @heathmont/moon-pagination@9.0.0
+  - @heathmont/moon-popover@9.0.0
+  - @heathmont/moon-select@9.0.0
+  - @heathmont/moon-sidebar@9.0.0
+  - @heathmont/moon-table@9.0.0
+  - @heathmont/moon-themes@9.0.0
+  - @heathmont/moon-utils@9.0.0
+  - @heathmont/moon-core-tw@9.0.0
+
+## 8.9.0
+
+### Minor Changes
+
+- TW config, moneyball theme and logo, image opts
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-accordion@8.9.0
+  - @heathmont/moon-assets@8.9.0
+  - @heathmont/moon-charts@8.9.0
+  - @heathmont/moon-components@8.9.0
+  - @heathmont/moon-core@8.9.0
+  - @heathmont/moon-datepicker@8.9.0
+  - @heathmont/moon-draggabletable@8.9.0
+  - @heathmont/moon-icons@8.9.0
+  - @heathmont/moon-pagination@8.9.0
+  - @heathmont/moon-popover@8.9.0
+  - @heathmont/moon-select@8.9.0
+  - @heathmont/moon-sidebar@8.9.0
+  - @heathmont/moon-table@8.9.0
+  - @heathmont/moon-themes@8.9.0
+  - @heathmont/moon-utils@8.9.0
+  - @heathmont/moon-core-tw@8.9.0
+
 ## 8.8.0
 
 ### Minor Changes