+++ packages/fonts/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Change Log
 
+## 9.0.1
+
+### Patch Changes
+
+- Bump corrupted 9.0.0
+
+## 9.0.0
+
+### Major Changes
+
+- React 18
+
+## 8.9.0
+
+### Minor Changes
+
+- TW config, moneyball theme and logo, image opts
+
 ## 8.8.0
 
 ### Minor Changes